pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "dfir-memdump"
version = "1.0.0"
description = "Memory forensics triage tool powered by Volatility3"
requires-python = ">=3.10"
dependencies = [
    "volatility3>=2.7.0",
    "yara-python>=4.5.0",
    "requests>=2.31.0",
    "click>=8.1.0",
    "rich>=13.7.0",
    "jinja2>=3.1.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
]

[project.scripts]
dfir-memdump = "dfir_memdump.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["dfir_memdump*"]

[tool.setuptools.package-data]
"*" = ["data/yara/*.yar", "templates/*.j2"]
